requirements.txt
packaging
triton
timm==0.4.12
pytest 
chardet 
yacs 
termcolor
submitit 
tensorboardX
fvcore
seaborn
scipy
matplotlib
opencv-python
transforms3d
tqdm
pyyaml
einops
kornia
kornia_moons
yacs
pillow
h5py
gdown
tables
vispy
glfw
albumentations
e2cnn
pydegensac